+++ config/config.php
@@ -70,7 +70,7 @@ $GLOBALS['TL_HOOKS']['processFormData'][]       = array('NotificationCenter\tl_f
 $GLOBALS['TL_HOOKS']['createNewUser'][]         = array('NotificationCenter\ContaoHelper', 'sendRegistrationEmail');
 $GLOBALS['TL_HOOKS']['updatePersonalData'][]    = array('NotificationCenter\ContaoHelper', 'sendPersonalDataEmail');
 $GLOBALS['TL_HOOKS']['getUserNavigation'][]     = array('NotificationCenter\ContaoHelper', 'addQueueToUserNavigation');
-
+$GLOBALS['TL_HOOKS']['activateAccount'][]       = array('NotificationCenter\ContaoHelper', 'sendActivationEmail');
 
 /**
  * Queue manager
@@ -100,8 +100,8 @@ $GLOBALS['NOTIFICATION_CENTER']['NOTIFICATION_TYPE'] = array_merge_recursive(
              'core_form' => array(
                  'recipients'           => array('admin_email', 'form_*', 'formconfig_*'),
                  'email_subject'        => array('form_*', 'formconfig_*', 'admin_email'),
-                 'email_text'           => array('form_*', 'formconfig_*', 'raw_data', 'admin_email'),
-                 'email_html'           => array('form_*', 'formconfig_*', 'raw_data', 'admin_email'),
+                 'email_text'           => array('form_*', 'formconfig_*', 'formlabel_*', 'raw_data', 'admin_email'),
+                 'email_html'           => array('form_*', 'formconfig_*', 'formlabel_*', 'raw_data', 'admin_email'),
                  'file_name'            => array('form_*', 'formconfig_*', 'admin_email'),
                  'file_content'         => array('form_*', 'formconfig_*', 'admin_email'),
                  'email_sender_name'    => array('admin_email', 'form_*', 'formconfig_*'),
@@ -110,6 +110,19 @@ $GLOBALS['NOTIFICATION_CENTER']['NOTIFICATION_TYPE'] = array_merge_recursive(
                  'email_recipient_bcc'  => array('admin_email', 'form_*', 'formconfig_*'),
                  'email_replyTo'        => array('admin_email', 'form_*', 'formconfig_*'),
                  'attachment_tokens'    => array('form_*', 'formconfig_*'),
+             ),
+             'member_activation' => array(
+                 'recipients'           => array('member_email', 'admin_email'),
+                 'email_subject'        => array('domain', 'member_*', 'admin_email'),
+                 'email_text'           => array('domain', 'member_*', 'admin_email'),
+                 'email_html'           => array('domain', 'member_*', 'admin_email'),
+                 'file_name'            => array('domain', 'member_*', 'admin_email'),
+                 'file_content'         => array('domain', 'member_*', 'admin_email'),
+                 'email_sender_name'    => array('admin_email', 'form_*'),
+                 'email_sender_address' => array('admin_email', 'form_*'),
+                 'email_recipient_cc'   => array('admin_email', 'member_*'),
+                 'email_recipient_bcc'  => array('admin_email', 'member_*'),
+                 'email_replyTo'        => array('admin_email', 'member_*'),
              ),
              'member_registration' => array(
                  'recipients'           => array('member_email', 'admin_email'),
@@ -126,10 +139,10 @@ $GLOBALS['NOTIFICATION_CENTER']['NOTIFICATION_TYPE'] = array_merge_recursive(
              ),
              'member_personaldata' => array(
                  'recipients'           => array('member_email', 'admin_email'),
-                 'email_subject'        => array('domain', 'member_*', 'member_old_*', 'recipient_email'),
-                 'email_text'           => array('domain', 'member_*', 'member_old_*', 'recipient_email'),
-                 'email_html'           => array('domain', 'member_*', 'member_old_*', 'recipient_email'),
-                 'email_sender_name'    => array('member_email', 'admin_email'),
+                 'email_subject'        => array('domain', 'member_*', 'member_old_*', 'changed_*', 'admin_email'),
+                 'email_text'           => array('domain', 'member_*', 'member_old_*', 'changed_*', 'admin_email'),
+                 'email_html'           => array('domain', 'member_*', 'member_old_*', 'changed_*', 'admin_email'),
+                 'email_sender_name'    => array('member_*'),
                  'email_sender_address' => array('member_email', 'admin_email'),
                  'email_recipient_cc'   => array('member_email', 'admin_email'),
                  'email_recipient_bcc'  => array('member_email', 'admin_email'),