+++ languages/en/tl_nc_notification.php
@@ -13,6 +13,7 @@
  */
 $GLOBALS['TL_LANG']['tl_nc_notification']['title']                   = array('Title', 'Please enter a title for this notification.');
 $GLOBALS['TL_LANG']['tl_nc_notification']['type']                    = array('Type', 'Please select a type for this notification.');
+$GLOBALS['TL_LANG']['tl_nc_notification']['flatten_delimiter']       = array('Delimiter for lists', 'When list values (array values) are submitted, the Notification Center flattens them automatically by using a comma (",") by default. Change this here if you like.');
 
 /**
  * Reference
@@ -41,6 +42,7 @@ $GLOBALS['TL_LANG']['tl_nc_notification']['config_legend']           = 'Configur
  */
 $GLOBALS['TL_LANG']['tl_nc_notification']['type']['contao']                 = 'Contao';
 $GLOBALS['TL_LANG']['tl_nc_notification']['type']['core_form']              = array('Form submission', 'This notification type can be sent when the form is submitted.');
+$GLOBALS['TL_LANG']['tl_nc_notification']['type']['member_activation']      = array('Member activation');
 $GLOBALS['TL_LANG']['tl_nc_notification']['type']['member_registration']    = array('Member registration');
 $GLOBALS['TL_LANG']['tl_nc_notification']['type']['member_personaldata']    = array('Member personal data');
 $GLOBALS['TL_LANG']['tl_nc_notification']['type']['member_password']        = array('Member lost password');