+++ library/NotificationCenter/ContaoHelper.php
@@ -21,23 +21,83 @@ class ContaoHelper extends \Controller
         parent::__construct();
     }
 
-
     /**
      * Send a registration e-mail
-     * @param integer
-     * @param array
-     * @param object
+     *
+     * @param int    $intId
+     * @param array  $arrData
+     * @param object $objModule
      */
-    public function sendRegistrationEmail($intId, $arrData, &$objModule)
+    public function sendRegistrationEmail($intId, $arrData, $objModule)
     {
         if (!$objModule->nc_notification) {
             return;
         }
 
-        $arrTokens                = array();
+        $arrTokens = array(
+            'link' => \Environment::get('base') . \Environment::get('request') . (($GLOBALS['TL_CONFIG']['disableAlias'] || strpos(\Environment::get('request'), '?') !== false) ? '&' : '?') . 'token=' . $arrData['activation']
+        );
+
+        // Disable the email to admin because no core notification has been sent
+        $objModule->reg_activate = true;
+
+        $this->sendNotifications($objModule->nc_notification, $arrData, $objModule, $arrTokens);
+    }
+
+    /**
+     * Send a activation e-mail
+     *
+     * @param object
+     * @param object
+     */
+    public function sendActivationEmail($objMember, &$objModule)
+    {
+        if (!$objModule->nc_activation_notification) {
+            return;
+        }
+
+        $this->sendNotifications($objModule->nc_activation_notification, $objMember->row(), $objModule);
+    }
+
+    /**
+     * Send the personal data change e-mail
+     *
+     * @param object $objUser
+     * @param array  $arrData
+     * @param object $objModule
+     */
+    public function sendPersonalDataEmail($objUser, $arrData, $objModule)
+    {
+        if (!$objModule->nc_notification) {
+            return;
+        }
+
+        $arrTokens = array();
+
+        // Translate/format old values
+        foreach ($_SESSION['PERSONAL_DATA'] as $strFieldName => $strFieldValue) {
+            $arrTokens['member_old_' . $strFieldName] = \Haste\Util\Format::dcaValue('tl_member', $strFieldName, $strFieldValue);
+        }
+
+        $this->sendNotifications($objModule->nc_notification, $arrData, $objModule, $arrTokens);
+    }
+
+    /**
+     * send the e-mail for the given module
+     *
+     * @param int
+     * @param array
+     * @param object
+     * @param array
+     */
+    private function sendNotifications($intNotification, $arrData, $objModule, $arrTokens = array())
+    {
+        if (!is_array($arrTokens)) {
+            $arrTokens = array();
+        }
+
         $arrTokens['admin_email'] = $GLOBALS['TL_ADMIN_EMAIL'];
         $arrTokens['domain']      = \Environment::get('host');
-        $arrTokens['link']        = \Environment::get('base') . \Environment::get('request') . (($GLOBALS['TL_CONFIG']['disableAlias'] || strpos(\Environment::get('request'), '?') !== false) ? '&' : '?') . 'token=' . $arrData['activation'];
 
         // Support newsletters
         if (in_array('newsletter', \ModuleLoader::getActive())) {
@@ -54,58 +114,15 @@ class ContaoHelper extends \Controller
         // translate/format values
         foreach ($arrData as $strFieldName => $strFieldValue) {
             $arrTokens['member_' . $strFieldName] = \Haste\Util\Format::dcaValue('tl_member', $strFieldName, $strFieldValue);
-        }
 
-        $objNotification = \NotificationCenter\Model\Notification::findByPk($objModule->nc_notification);
-
-        if ($objNotification !== null) {
-            $objNotification->send($arrTokens, $GLOBALS['TL_LANGUAGE']);
-
-            // Disable the email to admin because no core notification has been sent
-            $objModule->reg_activate = true;
-        }
-    }
-
-
-    /**
-     * Send the personal data change e-mail
-     * @param object
-     * @param array
-     * @param object
-     */
-    public function sendPersonalDataEmail($objUser, $arrData, $objModule)
-    {
-        if (!$objModule->nc_notification) {
-            return;
-        }
-
-        $arrTokens                = array();
-        $arrTokens['admin_email'] = $GLOBALS['TL_ADMIN_EMAIL'];
-        $arrTokens['domain']      = \Environment::get('host');
-
-        // Support newsletters
-        if (in_array('newsletter', $this->Config->getActiveModules())) {
-            if (!is_array($arrData['newsletter'])) {
-                if ($arrData['newsletter'] != '') {
-                    $objChannels                    = \Database::getInstance()->execute("SELECT title FROM tl_newsletter_channel WHERE id IN(" . implode(',', array_map('intval', (array) $arrData['newsletter'])) . ")");
-                    $arrTokens['member_newsletter'] = implode("\n", $objChannels->fetchEach('title'));
-                } else {
-                    $arrTokens['member_newsletter'] = '';
-                }
+            if ((string) $arrTokens['member_' . $strFieldName] !== (string) $arrTokens['member_old_' . $strFieldName]) {
+                $arrTokens['changed_' . $strFieldName] = '1';
+            } else {
+                $arrTokens['changed_' . $strFieldName] = '0';
             }
         }
 
-        // Translate/format old values
-        foreach ($_SESSION['PERSONAL_DATA'] as $strFieldName => $strFieldValue) {
-            $arrTokens['member_old_' . $strFieldName] = \Haste\Util\Format::dcaValue('tl_member', $strFieldName, $strFieldValue);
-        }
-
-        // Translate/format new values
-        foreach ($arrData as $strFieldName => $strFieldValue) {
-            $arrTokens['member_' . $strFieldName] = \Haste\Util\Format::dcaValue('tl_member', $strFieldName, $strFieldValue);
-        }
-
-        $objNotification = \NotificationCenter\Model\Notification::findByPk($objModule->nc_notification);
+        $objNotification = \NotificationCenter\Model\Notification::findByPk($intNotification);
 
         if ($objNotification !== null) {
             $objNotification->send($arrTokens, $GLOBALS['TL_LANGUAGE']);
@@ -115,8 +132,8 @@ class ContaoHelper extends \Controller
     /**
      * Remove Queue from back end navigation if no queue gateway is available yet
      *
-     * @param array
-     * @param bool
+     * @param array $arrModules
+     * @param bool  $blnShowAll
      *
      * @return array
      */
@@ -124,7 +141,6 @@ class ContaoHelper extends \Controller
     {
         // Make sure there's no exception if notification_center has not been properly installed yet
         if (!\Database::getInstance()->tableExists('tl_nc_gateway')) {
-
             return $arrModules;
         }