+++ library/NotificationCenter/Model/Language.php
@@ -10,6 +10,29 @@
 
 namespace NotificationCenter\Model;
 
+/**
+ * @property int    $id
+ * @property int    $pid
+ * @property int    $tstamp
+ * @property string $gateway_type
+ * @property string $language
+ * @property string $fallback
+ * @property string $recipients
+ * @property string $attachment_tokens
+ * @property string $attachments
+ * @property string $email_sender_name
+ * @property string $email_sender_address
+ * @property string $email_recipient_cc
+ * @property string $email_recipient_bcc
+ * @property string $email_replyTo
+ * @property string $email_subject
+ * @property string $email_mode
+ * @property string $email_text
+ * @property string $email_html
+ * @property string $file_name
+ * @property string $file_storage_mode
+ * @property string $file_content
+ */
 class Language extends \Model
 {