+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Notification Center Changelog
 ===========================
 
+Version 1.4.0 (2017-02-02)
+--------------------------
+
+### New
+- Added a verification for valid token characters (#68)
+- Support embedding images in e-mails (#109)
+- Support for different flatten delimiter in core form notification (#111)
+
+
 Version 1.3.6 (2016-05-04)
 --------------------------